Moving some cython functions into _utils.pyx
This moves a function from _loader/_loader.pyx and a function from
_scheduler/jobs/_job.pyx into _utils.pyx instead.

The reason for this change is that without it, the following commit
causes "Unable to import" pylint error E0401 to occur for job.py and
for loader.py respectively.

It is unclear what the following commit does to trigger these errors,
asides from adding some additional type safety from the typing module,
it affects mostly _frontend/cli.py, _context.py and _stream.py.

This change seems to be a good tradeoff however since the cython modules
each offer a single function that is not intertwined in the loader or
scheduler code at all, so it may even be better to just provide these in
the same soname (this might have load time benefits by reducing the number
of dynamic libraries to load).
