+++ .pylintrc
@@ -5,10 +5,8 @@
 # run arbitrary code
 extension-pkg-whitelist=
     buildstream.node,
-    buildstream._loader._loader,
     buildstream._loader.loadelement,
     buildstream._loader.types,
-    buildstream._scheduler.jobs._job,
     buildstream._types,
     buildstream._utils,
     buildstream._variables,
