--- src/buildstream/_loader/_loader.pyi
@@ -1,1 +1,0 @@
-def valid_chars_name(name: str) -> bool: ...
