--- src/buildstream/_scheduler/jobs/_job.pyi
@@ -1,1 +1,0 @@
-def terminate_thread(thread_id: int): ...
