+++ src/buildstream/_utils.pyi
@@ -1,1 +1,3 @@
 def url_directory_name(url: str) -> str: ...
+def valid_chars_name(name: str) -> bool: ...
+def terminate_thread(thread_id: int) -> None: ...
